Changed SwitCH Parameter, create ThemeMode Enum, changed some names...
