+++ app/src/main/java/com/example/space__chat/presentation/model/ThemeMode.kt
@@ -1,0 +1,6 @@
+package com.example.space__chat.Presentation.model
+
+enum class ThemeMode {
+    LIGHT,
+    DARK
+}
