+++ app/src/main/res/drawable/bckg_switch_dark.xml
