+++ app/src/main/res/drawable/bckg_switch_light.xml
